College Login System Update
This GUI-based program allows the user to input their login details to their associated college.

This includes the following:
1) Username
2) Password

If incorrect the program should return a message box error, else it will return that the login was successful.

There is the option to register as:
1) Student
2) Teacher
3) Admin

Each role permits different accessibilities...